Remove striping (see reasons in comment body)
When there are a lot of rows displayed, we wanted to show/hide rows with css, because it’s far more performance than adding and removing rows from the dom. Because we’re using css to show/hide rows, we couldn’t use :nth-child(odd) for striping. That left us with using javascript to do striping. Using javascript for striping can create a performance hit if we restripe the whole collection whenever a new row is shown. We can’t just keep track of add/even striping for newly added rows, because new shown rows aren’t guaranteed to show up in order that they are received; filtering and response timing can effect the order of which rows are shown. The only other solution is to use smart logic to restripe the current new visible row by looking at how the previous row was striped, and then also restripe any rows after. Doing so would require a lot of logic, and at this point it’s not worth the effort unless there’s a very strong desire for striping.